Separate out integration tests
