+++ pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+testpaths= tests
+python_files = tests.py test_*.py *_tests.py
+addopts = --strict --showlocals
